OMS Adapter: required changes
