+++ modules/omsBidAdapter.js
@@ -14,6 +14,7 @@ import {registerBidder} from '../src/adapters/bidderFactory.js';
 import {BANNER} from '../src/mediaTypes.js';
 import {config} from '../src/config.js';
 import {ajax} from '../src/ajax.js';
+import {percentInView} from "../libraries/percentInView/percentInView";
 
 const BIDDER_CODE = 'oms';
 const URL = 'https://rt.marphezis.com/hb';
@@ -21,6 +22,7 @@ const TRACK_EVENT_URL = 'https://rt.marphezis.com/prebid'
 
 export const spec = {
   code: BIDDER_CODE,
+  aliases: ['brightcom', 'bcmssp'],
   gvlid: 883,
   supportedMediaTypes: [BANNER],
   isBidRequestValid,
@@ -95,7 +97,7 @@ function buildRequests(bidReqs, bidderRequest) {
       deepSetValue(payload, 'regs.ext.us_privacy', bidderRequest.uspConsent);
     }
 
-    if (config.getConfig('coppa') === true) {
+    if (config.getConfig('coppa') === true || bidReqs[0].coppa) {
       deepSetValue(payload, 'regs.coppa', 1);
     }
 
@@ -226,7 +228,7 @@ function _isViewabilityMeasurable(element) {
 }
 
 function _getViewability(element, topWin, {w, h} = {}) {
-  return getWindowTop().document.visibilityState === 'visible' ? _getPercentInView(element, topWin, {w, h}) : 0;
+  return getWindowTop().document.visibilityState === 'visible' ? percentInView(element, topWin, {w, h}) : 0;
 }
 
 function _isIframe() {
@@ -241,68 +243,6 @@ function _getMinSize(sizes) {
   return sizes.reduce((min, size) => size.h * size.w < min.h * min.w ? size : min);
 }
 
-function _getBoundingBox(element, {w, h} = {}) {
-  let {width, height, left, top, right, bottom} = element.getBoundingClientRect();
-
-  if ((width === 0 || height === 0) && w && h) {
-    width = w;
-    height = h;
-    right = left + w;
-    bottom = top + h;
-  }
-
-  return {width, height, left, top, right, bottom};
-}
-
-function _getIntersectionOfRects(rects) {
-  const bbox = {
-    left: rects[0].left, right: rects[0].right, top: rects[0].top, bottom: rects[0].bottom
-  };
-
-  for (let i = 1; i < rects.length; ++i) {
-    bbox.left = Math.max(bbox.left, rects[i].left);
-    bbox.right = Math.min(bbox.right, rects[i].right);
-
-    if (bbox.left >= bbox.right) {
-      return null;
-    }
-
-    bbox.top = Math.max(bbox.top, rects[i].top);
-    bbox.bottom = Math.min(bbox.bottom, rects[i].bottom);
-
-    if (bbox.top >= bbox.bottom) {
-      return null;
-    }
-  }
-
-  bbox.width = bbox.right - bbox.left;
-  bbox.height = bbox.bottom - bbox.top;
-
-  return bbox;
-}
-
-function _getPercentInView(element, topWin, {w, h} = {}) {
-  const elementBoundingBox = _getBoundingBox(element, {w, h});
-
-  // Obtain the intersection of the element and the viewport
-  const elementInViewBoundingBox = _getIntersectionOfRects([{
-    left: 0, top: 0, right: topWin.innerWidth, bottom: topWin.innerHeight
-  }, elementBoundingBox]);
-
-  let elementInViewArea, elementTotalArea;
-
-  if (elementInViewBoundingBox !== null) {
-    // Some or all of the element is in view
-    elementInViewArea = elementInViewBoundingBox.width * elementInViewBoundingBox.height;
-    elementTotalArea = elementBoundingBox.width * elementBoundingBox.height;
-
-    return ((elementInViewArea / elementTotalArea) * 100);
-  }
-
-  // No overlap between element and the viewport; therefore, the element
-  // lies completely out of view
-  return 0;
-}
 
 function _getBidFloor(bid) {
   if (!isFn(bid.getFloor)) {
